Remove dead code and fix react-refresh warnings
- Delete classify-subject edge function (duplicate of generate-timeline
  mode='classify', never invoked from client)
- Delete SampleTimelineView + sampleEvents (mounted in App but never
  opened — setShowSampleTimeline(true) never called)
- Delete HelpPanel component (defined, never imported)
- Fix react-refresh/only-export-components warnings by splitting
  non-component exports: buttonVariants to its own file, useAuth and
  useSidePanel hooks to src/hooks/
