+++ src/components/Homepage/Homepage.tsx
@@ -1,6 +1,6 @@
 import { useState, useMemo, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '../../hooks/useAuth';
 import { useTimelines } from '../../hooks/useTimelines';
 import { useTimelineMetadata } from '../../hooks/useTimelineMetadata';
 import { supabase } from '../../lib/supabase';
