+++ src/components/Modal/ConfirmationModal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AlertTriangle } from 'lucide-react';
-import { buttonVariants } from '@/components/ui/button';
+import { buttonVariants } from '@/components/ui/buttonVariants';
 import {
   AlertDialog,
   AlertDialogAction,
