+++ src/components/SidePanel/GlobalSidePanel.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { MoreVertical, PanelLeft, Trash2, LogOut, Video } from 'lucide-react'
-import { useAuth } from '@/contexts/AuthContext'
-import { useSidePanel } from '@/contexts/SidePanelContext'
+import { useAuth } from '@/hooks/useAuth'
+import { useSidePanel } from '@/hooks/useSidePanel'
 import { useTimelines } from '@/hooks/useTimelines'
 import { supabase } from '@/lib/supabase'
 import { ConfirmationModal } from '@/components/Modal/ConfirmationModal'
