+++ src/components/ui/button.tsx
@@ -1,43 +1,9 @@
 import * as React from "react"
 import { Slot } from "@radix-ui/react-slot"
-import { cva, type VariantProps } from "class-variance-authority"
+import { type VariantProps } from "class-variance-authority"
 
 import { cn } from "@/lib/utils"
-
-const buttonVariants = cva(
-  "inline-flex shrink-0 items-center justify-center gap-2 rounded-md text-sm font-medium whitespace-nowrap transition-all outline-none focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 disabled:pointer-events-none disabled:opacity-50 aria-invalid:border-destructive aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 [&_svg]:pointer-events-none [&_svg]:shrink-0 [&_svg:not([class*='size-'])]:size-4",
-  {
-    variants: {
-      variant: {
-        default: "bg-primary text-primary-foreground hover:bg-primary/90",
-        destructive:
-          "bg-destructive text-white hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:bg-destructive/60 dark:focus-visible:ring-destructive/40",
-        outline:
-          "border bg-background shadow-sm hover:bg-accent hover:text-accent-foreground dark:border-input dark:bg-input/30 dark:hover:bg-input/50",
-        secondary:
-          "bg-secondary text-secondary-foreground hover:bg-secondary/80",
-        ghost:
-          "hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50",
-        link: "text-primary underline-offset-4 hover:underline",
-        glass:
-          "backdrop-blur-[12px] bg-white/10 border border-white/[0.15] rounded-xl h-[42px] min-w-[80px] gap-1.5 px-[17px] shadow-[0px_8px_32px_0px_rgba(0,0,0,0.4),inset_0px_1px_0px_0px_rgba(255,255,255,0.1)] text-[#c9ced4] hover:bg-white/20 hover:border-transparent hover:text-[#dadee5] hover:cursor-pointer active:bg-[rgba(37,99,235,0.4)] active:border-transparent active:text-[#dadee5] data-[active=true]:bg-[rgba(37,99,235,0.4)] data-[active=true]:border-transparent data-[active=true]:text-[#dadee5]",
-        "glass-sm":
-          "backdrop-blur-[12px] bg-white/10 border border-white/[0.15] rounded-[10px] min-w-[80px] gap-1 px-[11px] py-[6px] shadow-[0px_8px_32px_0px_rgba(0,0,0,0.4),inset_0px_1px_0px_0px_rgba(255,255,255,0.1)] text-[#c9ced4] hover:bg-white/20 hover:border-transparent hover:text-[#dadee5] hover:cursor-pointer active:bg-[rgba(37,99,235,0.4)] active:border-transparent active:text-[#dadee5] data-[active=true]:bg-[rgba(37,99,235,0.4)] data-[active=true]:border-transparent data-[active=true]:text-[#dadee5]",
-      },
-      size: {
-        default: "h-9 px-4 py-2 has-[>svg]:px-3",
-        sm: "h-8 gap-1.5 rounded-md px-3 has-[>svg]:px-2.5",
-        lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
-        icon: "size-9",
-        none: "",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-      size: "default",
-    },
-  }
-)
+import { buttonVariants } from "./buttonVariants"
 
 function Button({
   className,
@@ -62,4 +28,4 @@ function Button({
   )
 }
 
-export { Button, buttonVariants }
+export { Button }
